+++ simtools.gaml.extensions.physics/src/javax/vecmath/Tuple4i.java
@@ -351,10 +351,10 @@ public abstract class Tuple4i implements java.io.Serializable, Cloneable {
     @Override
     public int hashCode() {
 	long bits = 1L;
-	bits = 31L * bits + (long)x;
-	bits = 31L * bits + (long)y;
-	bits = 31L * bits + (long)z;
-	bits = 31L * bits + (long)w;
+	bits = 31L * bits + x;
+	bits = 31L * bits + y;
+	bits = 31L * bits + z;
+	bits = 31L * bits + w;
 	return (int) (bits ^ (bits >> 32));
     }
